# perf: Speed up trends queries involving user property breakdowns by 20x for instances with lots of people