--- a/ee/clickhouse/queries/person_distinct_id_query.py
+++ b/ee/clickhouse/queries/person_distinct_id_query.py
@@ -8,7 +8,7 @@ from posthog.settings import BENCHMARK, TEST
 using_new_table = TEST or BENCHMARK
 
 
-def get_team_distinct_ids_query(team_id: int) -> str:
+def get_team_distinct_ids_query(team_id: int, extra_where: str = "") -> str:
     from ee.clickhouse.client import substitute_params
 
     global using_new_table
@@ -16,9 +16,11 @@ def get_team_distinct_ids_query(team_id: int) -> str:
     using_new_table = using_new_table or _fetch_person_distinct_id2_ready()
 
     if using_new_table:
-        return substitute_params(GET_TEAM_PERSON_DISTINCT_IDS_NEW_TABLE, {"team_id": team_id})
+        return substitute_params(
+            GET_TEAM_PERSON_DISTINCT_IDS_NEW_TABLE.format(extra_where=extra_where), {"team_id": team_id,}
+        )
     else:
-        return substitute_params(GET_TEAM_PERSON_DISTINCT_IDS, {"team_id": team_id})
+        return substitute_params(GET_TEAM_PERSON_DISTINCT_IDS.format(extra_where=extra_where), {"team_id": team_id,})
 
 
 is_ready = False