--- a/ee/clickhouse/queries/retention/retention_event_query.py
+++ b/ee/clickhouse/queries/retention/retention_event_query.py
@@ -135,7 +135,12 @@ class RetentionEventsQuery(ClickhouseEventQuery):
         )
         self.params.update(entity_params)
 
-        person_query, person_params = self._get_person_query()
+        distinct_id_query, distinct_id_params = self._get_distinct_id_query(
+            entity_query=entity_query, date_query=date_query
+        )
+        self.params.update(distinct_id_params)
+
+        person_query, person_params = self._get_person_query(entity_query=entity_query, date_query=date_query)
         self.params.update(person_params)
 
         groups_query, groups_params = self._get_groups_query()
@@ -143,7 +148,7 @@ class RetentionEventsQuery(ClickhouseEventQuery):
 
         query = f"""
             SELECT {','.join(_fields)} FROM events {self.EVENT_TABLE_ALIAS}
-            {self._get_distinct_id_query()}
+            {distinct_id_query}
             {person_query}
             {groups_query}
             WHERE team_id = %(team_id)s