--- a/ee/clickhouse/queries/trends/trend_event_query.py
+++ b/ee/clickhouse/queries/trends/trend_event_query.py
@@ -64,7 +64,12 @@ class TrendsEventQuery(ClickhouseEventQuery):
         entity_query, entity_params = self._get_entity_query()
         self.params.update(entity_params)
 
-        person_query, person_params = self._get_person_query()
+        distinct_id_query, distinct_id_params = self._get_distinct_id_query(
+            entity_query=entity_query, date_query=date_query
+        )
+        self.params.update(distinct_id_params)
+
+        person_query, person_params = self._get_person_query(entity_query=entity_query, date_query=date_query)
         self.params.update(person_params)
 
         groups_query, groups_params = self._get_groups_query()
@@ -72,7 +77,7 @@ class TrendsEventQuery(ClickhouseEventQuery):
 
         query = f"""
             SELECT {_fields} FROM events {self.EVENT_TABLE_ALIAS}
-            {self._get_distinct_id_query()}
+            {distinct_id_query}
             {person_query}
             {groups_query}
             WHERE team_id = %(team_id)s