… ee/clickhouse/models/test/test_property.py
… ee/clickhouse/queries/trends/breakdown.py
… ee/clickhouse/queries/trends/lifecycle.py
… ee/clickhouse/sql/person.py