## Problem

Running PostHog Trends queries can be very slow when running a query:

- With user property breakdowns 
- Where the "Persons" table is extremely large (e.g., over 10 million users)

This was previously mentioned in https://github.com/PostHog/posthog/issues/7548, and while it's improved, it's still slow for tables with lots of people:

Here's an example query from our PostHog instance's screenshot:

![image](https://user-images.githubusercontent.com/2924388/159375976-fb1aea1b-9d59-4f94-b8b7-b1e7882aa596.png)

After investigating, the root cause is that Clickhouse is doing a bad job of optimizing the `person_distinct_id` query: instead of:

- Ideal query plan: doing the INNER JOIN on only the list of `distinct_ids` that might be relevant
- Actual query plan: doing the INNER JOIN on the whole (10 million-row) `persons` and `person_distinct_ids2` tables

For example, here's a pre-flight query from `get_breakdown_prop_values`. Our table has about 10 million people, and the relevant event happened about 1-10k times.

<details>
  <summary>Original query (13 seconds)</summary>
  
  ```sql
  SELECT
      replaceRegexpAll(JSONExtractRaw(person_props, 'signupPlatform'), concat('^[', regexpQuoteMeta('"'), ']*|[', regexpQuoteMeta('"'), ']*$'), '') AS value,
      sum(toFloat64OrNull(replaceRegexpAll(JSONExtractRaw(properties, 'revenue'), concat('^[', regexpQuoteMeta('"'), ']*|[', regexpQuoteMeta('"'), ']*$'), ''))) AS count
  FROM events AS e
  INNER JOIN
  (
      SELECT
          distinct_id,
          argMax(person_id, version) AS person_id
      FROM person_distinct_id2
      WHERE team_id = 2
      GROUP BY distinct_id
      HAVING argMax(is_deleted, version) = 0
  ) AS pdi ON e.distinct_id = pdi.distinct_id
  INNER JOIN
  (
      SELECT
          id,
          argMax(properties, _timestamp) AS person_props
      FROM person
      WHERE team_id = 2
      GROUP BY id
      HAVING max(is_deleted) = 0
  ) AS person ON pdi.person_id = person.id
  WHERE (team_id = 2) AND (event = 'Renew subscription on server') AND (timestamp >= '2021-12-16 00:00:00') AND (timestamp <= '2022-03-16 23:59:59')
  GROUP BY value
  ORDER BY count DESC
  LIMIT 0, 25

  Query id: 15aab7c2-a86a-4ed0-80d8-bd52fc50a4b0

  ┌─value───┬──────────────count─┐
  │ ios     │ 11593.386296337107 │
  │ android │  7643.970650707903 │
  │ web     │  5027.731255776903 │
  │         │     674.6225766581 │
  └─────────┴────────────────────┘

  4 rows in set. Elapsed: 13.421 sec. Processed 26.59 million rows, 3.80 GB (1.98 million rows/s., 282.84 MB/s.)
  ```
</details>

<details>
  <summary>New query with events filtering (0.7 seconds)</summary>

  ```sql
  SELECT
      replaceRegexpAll(JSONExtractRaw(person_props, 'signupPlatform'), concat('^[', regexpQuoteMeta('"'), ']*|[', regexpQuoteMeta('"'), ']*$'), '') AS value,
      sum(toFloat64OrNull(replaceRegexpAll(JSONExtractRaw(properties, 'revenue'), concat('^[', regexpQuoteMeta('"'), ']*|[', regexpQuoteMeta('"'), ']*$'), ''))) AS count
  FROM events AS e
  INNER JOIN
  (
      SELECT
          distinct_id,
          argMax(person_id, version) AS person_id
      FROM person_distinct_id2
      WHERE (team_id = 2) AND (distinct_id IN
      (
          SELECT distinct_id
          FROM events
          WHERE (team_id = 2) AND (event = 'Renew subscription on server') AND (timestamp >= '2021-12-16 00:00:00') AND (timestamp <= '2022-03-16 23:59:59')
          GROUP BY distinct_id
      ))
      GROUP BY distinct_id
      HAVING argMax(is_deleted, version) = 0
  ) AS pdi ON e.distinct_id = pdi.distinct_id
  INNER JOIN
  (
      SELECT
          id,
          argMax(properties, _timestamp) AS person_props
      FROM person
      WHERE (team_id = 2) AND (id IN
      (
          SELECT person_id
          FROM person_distinct_id2
          WHERE distinct_id IN
          (
              SELECT distinct_id
              FROM events
              WHERE (team_id = 2) AND (event = 'Renew subscription on server') AND (timestamp >= '2021-12-16 00:00:00') AND (timestamp <= '2022-03-16 23:59:59')
              GROUP BY distinct_id
          )
      ))
      GROUP BY id
      HAVING max(is_deleted) = 0
  ) AS person ON pdi.person_id = person.id
  WHERE (team_id = 2) AND (event = 'Renew subscription on server') AND (timestamp >= '2021-12-16 00:00:00') AND (timestamp <= '2022-03-16 23:59:59')
  GROUP BY value
  ORDER BY count DESC
  LIMIT 0, 25

  Query id: 133ee566-ca68-4a70-9f8b-db2412dbcdc6

  ┌─value───┬──────────────count─┐
  │ ios     │ 11593.386296337098 │
  │ android │  7643.970650707903 │
  │ web     │  5027.731255776903 │
  │         │  674.6225766581001 │
  └─────────┴────────────────────┘

  4 rows in set. Elapsed: 0.879 sec. Processed 12.08 million rows, 2.04 GB (13.75 million rows/s., 2.32 GB/s.)
  ```
</details>

The only differences above are that we pre-filter the rows we get back from the `persons` and `person_distinct_ids2` table with a subquery, like:

```sql
SELECT person_id
FROM person_distinct_id2
WHERE distinct_id IN
(
    SELECT distinct_id
    FROM events
    WHERE (team_id = 2) AND (event = 'Renew subscription on server') AND (timestamp >= '2021-12-16 00:00:00') AND (timestamp <= '2022-03-16 23:59:59')
    GROUP BY distinct_id
)
```

## Changes

Change our person_distinct_id2 (pdi) and persons (persons) table queries to use a subquery to help Clickhouse only retrieve the `persons` and `pdi` table entries relevant to the current query

- Specifically, we'll only get back people that have performed the events/actions in the timeframe given

## How did you test this code?

**I could use a lot of help here!** I tested the code only by running it on the local dev server and in my production PostHog cluster:

1. Following the steps in PostHog's development guide
2. Pushing it to my production cluster with the following Dockerfile:

Note: this was done using the `peter-release-1-33-0` branch: https://github.com/wanderlog/posthog/tree/peter-release-1-33-0

```docker
FROM posthog/posthog:release-1.33.0

COPY ee/clickhouse/queries/breakdown_props.py ee/clickhouse/queries/person_distinct_id_query.py ee/clickhouse/queries/person_query.py /home/posthog/code/ee/clickhouse/queries
COPY ee/clickhouse/queries/trends/breakdown.py /home/posthog/code/ee/clickhouse/queries/trends
COPY ee/clickhouse/queries/funnels/funnel_correlation.py /home/posthog/code/ee/clickhouse/queries/funnels
COPY ee/clickhouse/sql/person.py /home/posthog/code/ee/clickhouse/sql
```

I ran:

```sh
docker pull posthog/posthog:release-1.33.0 && docker build --tag=wanderlog/posthog:release-1.33.0 . && docker push wanderlog/posthog:release-1.33.0
```

However, this could definitely use automated tests -- could someone point me to the right places to add them?