@pxpeterxu 2022-03-24
Just wanted to leave a paper-trail -- I chatted with Karl, and we verified the performance improvement! That being said, it seems like this workstream is a bit at odds with the improvements at https://github.com/PostHog/posthog/issues/8132, which should fix this once and for all.

I'm also totally fine with not merging this in and just waiting for that change